fix(jit): re-intern PrimCall strings across dual heaps (#2575)
Evaluator prims allocate on string_heap_ while JIT PrimDisplay uses
g_string_pool under the same v2 encoding. aura_jit_prim_dispatch now
converts string args JIT→eval and re-interns results eval→JIT via
aura_alloc_string so string-append / number->string display correctly.
Also align aura_display_value write_mode ABI to int64_t.

+++ src/compiler/aura_jit.cpp
@@ -2454,7 +2454,7 @@ uint64_t aura_prim_call_count();
 uint64_t aura_prim_call_total_ns();
 void aura_display_int(int64_t);
 void aura_display_char(char);
-void aura_display_value(int64_t, int); // Issue #2572
+void aura_display_value(int64_t, int64_t); // Issue #2572 / #2575 ABI i64
 void aura_newline();
 void aura_jit_epoch_acquire_fence(void);
 void aura_jit_linear_post_invalidate_safety(std::uint8_t linear_state, std::uint32_t opcode);
